Completion of Part 5: Ride Cycle Functionality
The runOneCycle method within the Ride class was corrected to include functions like checking if the waiting line is empty, making sure there is an operator assigned to the ride and checking if there is more than 1 rider available. This method removes the 3 most current riders (from top) from the waiting line ArrayList and adds them to the ride. This represents an actual roller coaster (if there were only one carriage). These methods were called within the partFive section, in the AssignmentTwo.java.